Added upgrade queue
