+++ src/wazuh_modules/agent_upgrade/manager/wm_agent_upgrade_tasks.c
@@ -161,11 +161,11 @@ void wm_agent_upgrade_insert_task_id(int agent_id, int task_id) {
     }
 }
 
-void wm_agent_upgrade_remove_entry(int agent_id) {
+void wm_agent_upgrade_remove_entry(int agent_id, int free) {
     char agent_id_string[128];
     sprintf(agent_id_string, "%d", agent_id);
     wm_agent_task *agent_task = (wm_agent_task *)OSHash_Delete_ex(task_table_by_agent_id, agent_id_string);
-    if (agent_task) {
+    if (free) {
         wm_agent_upgrade_free_agent_task(agent_task);
     }
 }
