+++ src/wazuh_modules/agent_upgrade/manager/wm_agent_upgrade_tasks.h
@@ -112,8 +112,9 @@ void wm_agent_upgrade_insert_task_id(int agent_id, int task_id);
 /**
  * Removes an entry based on the agent_id
  * @param agent_id id of the agent
+ * @param free whether free task or not
  * */
-void wm_agent_upgrade_remove_entry(int agent_id);
+void wm_agent_upgrade_remove_entry(int agent_id, int free);
 
 /**
  * Returns the first node of the tasks hash table
@@ -167,7 +168,7 @@ cJSON* wm_agent_upgrade_send_tasks_information(const cJSON *message_object) __at
  * @retval OS_INVALID on errors
  * 
  * */
-int wm_agent_upgrade_task_module_callback(cJSON *json_response, const cJSON* task_module_request, cJSON* (*success_callback)(int *error, cJSON* input_json), void (*error_callback)(int agent_id)) __attribute__((nonnull(1,2)));
+int wm_agent_upgrade_task_module_callback(cJSON *json_response, const cJSON* task_module_request, cJSON* (*success_callback)(int *error, cJSON* input_json), void (*error_callback)(int agent_id, int free)) __attribute__((nonnull(1,2)));
 
 /**
  * Callback defined for upgrade command to process task manager information reponse
